+++ src/main/java/roomescape/auth/AdminInterceptor.java
@@ -4,7 +4,7 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.member.Member;
+import roomescape.member.domain.Member;
 import roomescape.member.MemberService;
 
 public class AdminInterceptor implements HandlerInterceptor {
