+++ src/main/java/roomescape/auth/AuthArgumentResolver.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.member.Member;
+import roomescape.member.domain.Member;
 import roomescape.member.MemberService;
 
 public class AuthArgumentResolver implements HandlerMethodArgumentResolver {
