+++ src/main/java/roomescape/auth/JwtTokenProvider.java
@@ -1,6 +1,5 @@
 package roomescape.auth;
 
-import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import java.security.Key;
