+++ src/main/java/roomescape/auth/dto/LoginCheckResponse.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.auth.dto;
 
 public class LoginCheckResponse {
