+++ src/main/java/roomescape/auth/dto/LoginRequest.java
@@ -1,13 +1,9 @@
-package roomescape.member;
+package roomescape.auth.dto;
 
-public class MemberRequest {
-    private String name;
+public class LoginRequest {
+
     private String email;
     private String password;
-
-    public String getName() {
-        return name;
-    }
 
     public String getEmail() {
         return email;
