+++ src/main/java/roomescape/member/LoginMemberArgumentResolver.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+import roomescape.member.domain.Member;
 
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
