+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -11,8 +11,10 @@ import org.springframework.web.bind.annotation.RestController;
 import java.net.URI;
 import java.util.List;
 import roomescape.auth.LoginMember;
-import roomescape.member.Member;
+import roomescape.member.domain.Member;
 import roomescape.member.MemberService;
+import roomescape.reservation.dto.ReservationRequest;
+import roomescape.reservation.dto.ReservationResponse;
 
 @RestController
 public class ReservationController {
