+++ src/main/java/roomescape/theme/ThemeController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URI;
 import java.util.List;
+import roomescape.theme.domain.Theme;
 
 @RestController
 public class ThemeController {
