+++ src/main/java/roomescape/time/TimeService.java
@@ -1,8 +1,10 @@
 package roomescape.time;
 
 import org.springframework.stereotype.Service;
-import roomescape.reservation.Reservation;
 import roomescape.reservation.ReservationDao;
+import roomescape.reservation.domain.Reservation;
+import roomescape.time.domain.AvailableTime;
+import roomescape.time.domain.Time;
 
 import java.util.List;
